Cache versions from changelog of each modified source package using JSON.
This will speed up the comparison process by skipping the unpacking of
many source packages, which is an expensive process. The parsing of the
changelogs is also skipped, but that wasn't much of a speed concern.

For example: Aptosid went from 56s on the first run to 4s on later runs.