Switch from the current hacky apt config method to using APT_CONFIG.
APT_CONFIG is the correct way to use an entirely separate set of
apt configuration and data directories.

This isolates census apt from the system apt config, which could
include hooks and other stuff that shouldn't be run by the census.

This reduces the amount of apt configuration duplication within the
census scripts so only one place needs to be changed.

This allows the addition of extra per-derivative apt configuration
settings and sources.list entries.